+++ kamehouse-shell/bin/deploy/deploy-kamehouse-dev.sh
@@ -46,7 +46,7 @@ deployKameHouseMobile() {
 
 # Get kamehouse httpd content root directory
 getHttpdContentRoot() {
-  if ${IS_LINUX_HOST}; then
+  if ${IS_LIN_HOST}; then
     echo "/var/www/kamehouse-webserver-dev"  
   else
     echo "${HOME}/programs/apache-httpd/www/kamehouse-webserver-dev"