+++ kamehouse-shell/bin/deploy/set-httpd-dir.sh
@@ -1,7 +1,7 @@
 HTTPD_DIR="${HOME}/programs/apache-httpd"
 
 setHttpDir() {
-  if ${IS_LINUX_HOST}; then
+  if ${IS_LIN_HOST}; then
     HTTPD_DIR="/etc/apache2"
   fi  
   export HTTPD_DIR="${HTTPD_DIR}"