+++ kamehouse-shell/bin/docker/docker-ci-integration-tests-trigger.sh
@@ -110,7 +110,7 @@ killCiContainerProcesses() {
 }
 
 killProcessRunningOnPort() {
-  if ! ${IS_LINUX_HOST}; then
+  if ! ${IS_LIN_HOST}; then
     return
   fi
   local PORT=$1