+++ kamehouse-shell/bin/functions/vlc/vlc-functions.sh
@@ -54,7 +54,7 @@ removeVlcProcessInfo() {
 
 searchForActiveVlcProcess() {
   log.debug "Searching for vlc process with an http server"
-  if ${IS_LINUX_HOST}; then
+  if ${IS_LIN_HOST}; then
     VLC_PID=`netstat -nltp 2>/dev/null | grep ":${VLC_PORT} " | grep vlc | grep -v tcp6 | awk '{print $7}' | cut -d '/' -f 1`
   else
     VLC_PID=`netstat -ano | grep "LISTENING" | grep "\[::\]:${VLC_PORT} " | tail -n 1 | awk '{print $5}' | cut -d '/' -f 1`